pyproject.toml
[project]
name = "search-orchestrator"
version = "0.1.0"
requires-python = ">=3.12"

dependencies = [
    "httpx",
    "beautifulsoup4",
    "pyyaml",
    "redis",
    "fastapi",
    "uvicorn",
    "pyjwt",
    "sqlalchemy>=2.0",
    "aiosqlite",
    "bcrypt",
    "pydantic[email]",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "pytest-asyncio",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["adapters*", "normalizer*", "ranker*", "cache*", "api*", "domain*", "auth*", "alerts*"]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
